% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEAD.R
\name{get_ci}
\alias{get_ci}
\title{Get the confidence interval of cell type proportions}
\usage{
get_ci(mead.out, alpha = 0.05)
}
\arguments{
\item{mead.out}{output from running MEAD}

\item{alpha}{(1-alpha) confidence interval}
}
\value{
a list of two matrices: ci_l and ci_r, corresponding to lower and upper bound of the confidence intervals.
}
\description{
Get the confidence interval of cell type proportions
}
