% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_func.R
\name{get_two_group_diff}
\alias{get_two_group_diff}
\title{return the estimated group mean difference}
\usage{
get_two_group_diff(groups, p_hat)
}
\description{
return the estimated group mean difference
}
