% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_func.R
\name{J_softplus}
\alias{J_softplus}
\title{Jacobian matrix of softplus scale function}
\usage{
J_softplus(b, a)
}
\arguments{
\item{b}{beta_hat}

\item{a}{control param}
}
\value{
Jacobian matrix
}
\description{
Jacobian matrix of softplus scale function
}
