% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_func.R
\name{J_sum2one}
\alias{J_sum2one}
\title{Jacobian matrix of sum-to-1 scale function}
\usage{
J_sum2one(b, K)
}
\arguments{
\item{b}{beta_hat}

\item{K}{number of cell types}
}
\value{
Jacobian matrix
}
\description{
Jacobian matrix of sum-to-1 scale function
}
