% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEAD.R
\name{MEAD}
\alias{MEAD}
\title{Preprocess bulk and reference datasets}
\usage{
MEAD(
  bulk,
  ref,
  cell_types = NULL,
  preprocessing_control = list(),
  estimation_control = list(),
  verbose = FALSE,
  R01 = NULL
)
}
\arguments{
\item{bulk}{bulk samples to be deconvolved. A count matrix, whose rownames are gene names and colnames are individual names.}

\item{ref}{reference dataset, a SingleCellExperiment object, with phenotype: cell_type, individual}

\item{cell_types}{cell types to be used in the deconvolution. If NULL, will use the cell types in the ref dataset.}

\item{preprocessing_control}{a list of control parameters passed to MEAD_preprocessing}

\item{estimation_control}{a list of control parameters passed to MEAD_est}
}
\value{
a list of:
 *p_hat: estimated cell type proportions
 *p_hat_se: standard errors of p_hat
 *p_group_diff_hat: estimated two group means different if groups labels are provided.(levels(as.factor(groups)))
}
\description{
Preprocess bulk and reference datasets
}
