% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_func.R
\name{MEAD_est}
\alias{MEAD_est}
\title{Function performs estimation and inference.}
\usage{
MEAD_est(
  y,
  X,
  Vg,
  w = NULL,
  hc.type = "hc3",
  centeringXY = FALSE,
  nfold = 10,
  folds = NULL,
  R01 = NULL,
  groups = NULL,
  calc_var = TRUE,
  use.QP = FALSE,
  lambda = NULL,
  beta_tilde_transformation = list(method = "truncate", a = 10)
)
}
\arguments{
\item{y}{a vector of bulk sample}

\item{X}{reference matrix, estimated relative expression, gene by cell}

\item{Vg}{variance matrix of X, either G by K or G by K^2.}

\item{w}{gene weights; if null, w = 1}

\item{hc.type}{'hc3', 'hc2', 'hc0', 'cv' or 'cv_indep'}

\item{nfold}{if using cv, the number of fold to be used. If folds not provided, will run k-medoids on 11' - R01 to determine folds.}

\item{folds}{if using cv, the folds to be used. folds looks like 111222333.}

\item{R01}{the 0-1 correlation matrix}

\item{calc_var}{whether calculate the covariance matrices.}

\item{use.QP}{whether use quadratic programming to impose nonnegative constraint}
}
\value{
a list including: 1. p_hat: estimated cell type proportions; 2. p_hat_se: their standard errors; 3. p_group_diff_hat: if groups provided, estimated group mean difference
}
\description{
Function performs estimation and inference.
}
