% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{MEAD_preprocessing}
\alias{MEAD_preprocessing}
\title{Preprocess bulk and reference datasets}
\usage{
MEAD_preprocessing(
  bulk,
  ref,
  marker_gene = NULL,
  cell_types = NULL,
  gene_thresh = 0.05,
  max_count_quantile_celltype = 0.99,
  max_count_quantile_indi = 0.99,
  filter.gene = TRUE
)
}
\arguments{
\item{bulk}{bulk samples to be deconvolved. A count matrix, whose rownames are gene names and colnames are individual names.}

\item{ref}{reference dataset, a SingleCellExperiment object, with phenotype: cell_type, individual}

\item{marker_gene}{marker genes to use if any. Default is NULL.}

\item{cell_types}{cell types to be used in the deconvolution. If NULL, will use the cell types in the ref dataset.}

\item{gene_thresh}{a fraction. Remove genes that have expression in less than (gene_thresh * #cells) cells(less than 5% of total cells by default)}

\item{max_count_quantile_celltype}{a fraction. Remove union of genes that expressed more than max_count_quantile_celltype in each cell type.}

\item{max_count_quantile_indi}{a fraction. Remove union of genes that expressed more than max_count_quantile_indi in each individual.}

\item{verbose}{whether print progress}
}
\value{
a list of:
 *bulk: bulk samples
 *ref: ref samples
 *cell_types: cell_types to be considered
 *individuals: individuals in the reference dataset
}
\description{
Preprocess bulk and reference datasets
}
