% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_hc3.R
\name{ols_hc3}
\alias{ols_hc3}
\title{Fit ols and use hc3 for variance estimation}
\usage{
ols_hc3(y, X, w = NULL, groups = NULL, calc_var = TRUE)
}
\arguments{
\item{y, X}{response matrix and covariate matrix}

\item{w}{weights}

\item{groups}{a vector of group indicators, indicating the group of each column of y. For example, c(1,1,2,2).}

\item{calc_var}{whether perform variance estimation}
}
\value{
a list of:
 *p_hat: estimated cell type proportions
 *p_hat_se: standard errors of p_hat
 *p_group_diff_hat: estimated two group means different if groups labels are provided.(levels(as.factor(groups)))
 *p_group_diff_hat_se: estimated se of two group means different if groups labels are provided.
}
\description{
Fit ols and use hc3 for variance estimation
}
