% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEAD.R
\name{preprocessing_control_default}
\alias{preprocessing_control_default}
\title{Default MEAD_preprocessing input parameters}
\usage{
preprocessing_control_default()
}
\description{
Default MEAD_preprocessing input parameters
}
